Django/requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.6.3
django-oauth2-provider==0.2.6.1
django-rest-swagger==0.3.3
django-serializers==0.6.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
djangorestframework-gis==0.9.3
gunicorn==19.3.0
jsonfield==1.0.3
multigtfs==0.4.3
protobuf==3.0.0a4
psycopg2==2.6.1
PyYAML==3.11
requests==2.7.0
shortuuid==0.4.2
six==1.9.0
South==1.0.2
static3==0.6.1
wheel==0.24.0
